Make swaggerSchema an option and remove types file
