+++ src/swagger.ts
@@ -1,19 +1,13 @@
+import { JSONSchemaType, JSONSchemaTypes, isObjectType } from './json-schema';
 import {
-  Endpoint,
   EndpointParam,
-  Endpoints,
-  GraphQLParameters,
-  isOa3Param,
-  JSONSchemaType,
-  OA3BodyParam,
-  Oa3Param,
-  OperationObject,
-  Param,
-  Responses,
-  SwaggerSchema,
-} from './types';
-import { getRequestOptions } from './getRequestOptions';
-import { isObjectType } from './json-schema';
+  getRequestOptions,
+  RequestOptions,
+} from './getRequestOptions';
+
+export interface GraphQLParameters {
+  [key: string]: any;
+}
 
 const replaceOddChars = (str: string): string =>
   str.replace(/[^_a-zA-Z0-9]/g, '_');
@@ -22,6 +16,16 @@ const getGQLTypeNameFromURL = (method: string, url: string): string => {
   const fromUrl = replaceOddChars(url.replace(/[{}]+/g, ''));
   return `${method}${fromUrl}`;
 };
+
+export interface Responses {
+  [key: string]: {
+    schema?: JSONSchemaType;
+    content?: {
+      'application/json': { schema: JSONSchemaType };
+    };
+    type?: 'file';
+  };
+}
 
 export const getSuccessResponse = (
   responses: Responses,
@@ -49,6 +53,47 @@ export const getSuccessResponse = (
   return undefined;
 };
 
+export interface BodyParam {
+  name: string;
+  required?: boolean;
+  schema: JSONSchemaType;
+  in: 'body';
+}
+
+export interface Oa2NonBodyParam {
+  name: string;
+  type: JSONSchemaTypes;
+  in: 'header' | 'query' | 'formData' | 'path';
+  required?: boolean;
+}
+
+export interface Oa3Param {
+  name: string;
+  in: 'header' | 'query' | 'formData' | 'path';
+  required?: boolean;
+  schema: JSONSchemaType;
+}
+
+export type NonBodyParam = Oa2NonBodyParam | Oa3Param;
+
+export type Param = BodyParam | NonBodyParam;
+
+export interface OA3BodyParam {
+  content: {
+    'application/json'?: {
+      schema: JSONSchemaType;
+    };
+    'application/x-www-form-urlencoded'?: {
+      schema: JSONSchemaType;
+    };
+  };
+  description?: string;
+  required: boolean;
+}
+
+export const isOa3Param = (param: Param): param is Oa3Param => {
+  return !!(param as Oa3Param).schema;
+};
 
 export function addTitlesToJsonSchemas(schema: SwaggerSchema): SwaggerSchema {
   const requestBodies = (schema.components || {}).requestBodies || {};
@@ -185,6 +230,67 @@ export const getParamDetailsFromRequestBody = (
 
 function isFormdataRequest(requestBody: OA3BodyParam): boolean {
   return !!requestBody.content[contentTypeFormData];
+}
+
+export interface Endpoint {
+  parameters: EndpointParam[];
+  description?: string;
+  response: JSONSchemaType | undefined;
+  getRequestOptions: (args: GraphQLParameters) => RequestOptions;
+  mutation: boolean;
+}
+
+export interface Endpoints {
+  [operationId: string]: Endpoint;
+}
+
+export interface OperationObject {
+  requestBody?: OA3BodyParam;
+  description?: string;
+  operationId?: string;
+  parameters?: Param[];
+  responses: Responses;
+  consumes?: string[];
+}
+
+export type PathObject = {
+  parameters?: Param[];
+} & {
+  [operation: string]: OperationObject;
+};
+
+export interface Variable {
+  default?: string;
+  enum: string[];
+}
+
+export interface ServerObject {
+  url: string;
+  description?: string;
+  variables: {
+    [key: string]: string | Variable;
+  };
+}
+
+export interface SwaggerSchema {
+  host?: string;
+  basePath?: string;
+  schemes?: [string];
+  servers?: ServerObject[];
+  paths: {
+    [pathUrl: string]: PathObject;
+  };
+  components?: {
+    requestBodies?: {
+      [name: string]: OA3BodyParam;
+    };
+    schemas?: {
+      [name: string]: JSONSchemaType;
+    };
+  };
+  definitions?: {
+    [name: string]: JSONSchemaType;
+  };
 }
 
 /**
