+++ test/fixture-test.ts
@@ -11,7 +11,8 @@ describe('Fixture', () => {
       describe(file, () => {
         const graphqlFile = file.replace('.json', '.graphql');
         it(`should convert to ${graphqlFile}`, () =>
-          graphQLSchema(directory + file, {
+          graphQLSchema({
+            swaggerSchema: directory + file,
             callBackend() {
               return new Promise(() => {});
             },
@@ -27,9 +28,10 @@ describe('Fixture', () => {
 
   describe('petstore converted to openapi 3', () => {
     it('should have the same graphql schema as openapi 2', async () => {
-      const file = `test/fixtures/petstore-openapi3.yaml`;
+      const swaggerSchema = `test/fixtures/petstore-openapi3.yaml`;
       const graphqlFile = `test/fixtures/petstore.graphql`;
-      const schema = await graphQLSchema(file, {
+      const schema = await graphQLSchema({
+        swaggerSchema,
         callBackend() {
           return new Promise(() => {});
         },
